refactor(SiteNavigation): update app name and improve user guide
Update the app name in the SiteNavigation component from "本地天气" to "WeatherApp". This change improves the clarity and consistency of the app.

Improve the user guide in the About section of the SiteNavigation component by making the instructions more general. Instead of referring to "your city", it now refers to "the selected city". This change makes the user guide applicable to any city selected by the user.

Also, update the instructions in the user guide for deleting a city. Instead of selecting the city on the homepage and finding the delete option at the bottom of the page, the user can now delete a city by clicking the trash can icon in the banner.

Finally, add information about navigating to the homepage by clicking the top-left corner and add version information in the Version Info section.

Issue: main